refactor: token 애플리케이션 계층 리팩터링
