+++ src/main/java/com/example/demo/application/token/api/TokenApi.java
@@ -1,8 +1,7 @@
-package com.example.demo.domain.token.api;
+package com.example.demo.application.token.api;
 
-import com.example.demo.domain.token.domain.dto.TokenRequest;
-import com.example.demo.domain.token.domain.dto.TokenResponse;
-import com.example.demo.domain.user.domain.dto.response.UserProfile;
+import com.example.demo.application.token.dto.TokenRequest;
+import com.example.demo.application.token.dto.TokenResponse;
 import com.example.demo.global.base.dto.ResponseBody;
 import com.example.demo.global.base.exception.ErrorCode;
 import com.example.demo.global.config.swagger.ApiErrorResponseExplanation;
