+++ src/main/java/com/example/demo/application/token/dto/TokenRequest.java
@@ -1,4 +1,4 @@
-package com.example.demo.domain.token.domain.dto;
+package com.example.demo.application.token.dto;
 
 import jakarta.validation.constraints.NotBlank;
